# HEX-LoadBench/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HEX-LoadBench API
  description: |
    Load-Testing & API Stress Test Platform
    
    ## WARNING
    THIS TOOL IS FOR AUTHORIZED PERFORMANCE TESTING ONLY.
    
    Users are responsible for ensuring they have proper authorization before testing any systems.
    
    ## Authentication
    This API uses JWT tokens and API keys for authentication.
    
    - JWT tokens are used for user authentication (session-based)
    - API keys are used for service-to-service communication
    - All endpoints require authentication except `/health` and `/docs`
    
  version: 1.0.0
  contact:
    name: HEX-LoadBench Support
    email: support@hexloadbench.com
    url: https://hexloadbench.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.hexloadbench.com
    description: Production server
paths:
...
  /plans/subscriptions:
    get:
      tags:
        - Plans
      summary: List subscription plans
      description: Returns available subscription plans
      operationId: listSubscriptionPlans
      security:
        - bearerAuth: []
      parameters:
        - name: public_only
          in: query
          description: Return only public plans
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Plans retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionPlanResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/my-subscription:
    get:
      tags:
        - Plans
      summary: Get current user subscription
      description: Returns the current user's subscription information
      operationId: getMySubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSubscriptionWithPlan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No active subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /plans/usage:
    get:
      tags:
        - Plans
      summary: Get usage report
      description: Returns current usage and limits for the user
      operationId: getUsageReport
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage report retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageReport'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: number
          format: float
          description: Unix timestamp
...
    UserSubscriptionWithPlan:
      type: object
      properties:
        id:
          type: integer
          description: Subscription ID
        user_id:
          type: integer
          description: User ID
        plan_id:
          type: string
          description: Plan identifier
        status:
          type: string
          enum: [active, expired, cancelled, suspended]
          description: Subscription status
        start_date:
          type: string
          format: date-time
          description: Subscription start date
        end_date:
          type: string
          format: date-time
          description: Subscription end date
        billing_cycle:
          type: string
          enum: [monthly, yearly]
          description: Billing cycle
        current_daily_rps:
          type: integer
          description: Current daily RPS usage
        current_concurrent_tests:
          type: integer
          description: Current concurrent tests
        plan:
          $ref: '#/components/schemas/SubscriptionPlanResponse'
    SubscriptionPlanResponse:
      type: object
      properties:
        id:
          type: integer
          description: Plan ID
        plan_id:
          type: string
          description: Plan identifier
        name:
          type: string
          description: Plan name
        description:
          type: string
          description: Plan description
        max_daily_rps:
          type: integer
          description: Maximum daily RPS allowed
        max_concurrent_tests:
          type: integer
          description: Maximum concurrent tests allowed
        max_test_duration:
          type: integer
          description: Maximum test duration in seconds
        max_users_per_organization:
          type: integer
          description: Maximum users per organization
        max_api_keys:
          type: integer
          description: Maximum API keys allowed
        monthly_cost:
          type: number
          format: float
          description: Monthly cost
        is_active:
          type: boolean
          description: Whether the plan is active
        is_public:
          type: boolean
          description: Whether the plan is publicly available
...
    UsageLimits:
      type: object
      properties:
        max_daily_rps:
          type: integer
          description: Maximum daily RPS
        max_concurrent_tests:
          type: integer
          description: Maximum concurrent tests
        max_test_duration:
          type: integer
          description: Maximum test duration
        max_users_per_organization:
          type: integer
          description: Maximum users per organization
        max_api_keys:
          type: integer
          description: Maximum API keys
    CurrentUsage:
      type: object
      properties:
        daily_rps:
          type: integer
          description: Current daily RPS
        concurrent_tests:
          type: integer
          description: Current concurrent tests
        test_duration:
          type: integer
          description: Current test duration
        users_in_organization:
          type: integer
          description: Current users in organization
        api_keys_active:
          type: integer
          description: Active API keys
    UsageReport:
      type: object
      properties:
        limits:
          $ref: '#/components/schemas/UsageLimits'
        current:
          $ref: '#/components/schemas/CurrentUsage'
        remaining:
          $ref: '#/components/schemas/UsageLimits'
        usage_percentage:
          type: object
          additionalProperties:
            type: number
          description: Usage percentages for each metric
...